Fix: cardId recoil atom 수정

+++ src/components/pages/v3/V3OrgCardDetailPage/InformationPanel/index.tsx
@@ -1,10 +1,10 @@
 import React, {memo} from 'react';
-import {useRecoilValue} from 'recoil';
+import {useRecoilState, useRecoilValue, useSetRecoilState} from 'recoil';
 import {MobileSection} from '^v3/share/sections/MobileSection';
 import {AiOutlineEdit} from 'react-icons/ai';
 import {useModal} from '../../share/modals/useModal';
 import {currentCreditCardAtom} from '../../V3OrgCardListPage/modals/atom';
-import {cardIdParamState, orgIdParamState} from '^atoms/common';
+import {orgIdParamState} from '^atoms/common';
 import {creditCardApi} from '^api/credit-cards.api';
 import {useRouter} from 'next/router';
 import {V3OrgCardListPageRoute} from '^pages/v3/orgs/[orgId]/cards';
@@ -12,7 +12,7 @@ import {inputCardNumberModal} from '../../V3OrgCardListPage/modals/CardNumberMod
 import {selectCardCompanyModal} from '../../V3OrgCardListPage/modals/CardCompanyModal/atom';
 import {inputCardNameModal} from '../../V3OrgCardListPage/modals/CardNameModal/atom';
 import {inputCardHoldingMemberModal} from '../../V3OrgCardListPage/modals/CardHoldingMemberModal/atom';
-import {creditCardSignAtom} from '^models/CreditCard/atom';
+import {cardIdParamState, creditCardSignAtom} from '^models/CreditCard/atom';
 import {useAlert} from '^hooks/useAlert';
 
 export const InformationPanel = memo(() => {
@@ -23,18 +23,21 @@ export const InformationPanel = memo(() => {
     const {open: openInputCardHoldingMemberModal} = useModal(inputCardHoldingMemberModal);
     const {open: openSelectCardCompanyModal} = useModal(selectCardCompanyModal);
     const orgId = useRecoilValue(orgIdParamState);
-    const cardId = useRecoilValue(cardIdParamState);
+    const [cardId, setCardId] = useRecoilState(cardIdParamState);
     const router = useRouter();
     const {alert} = useAlert();
 
     // 카드 삭제 함수
     const onDelete = () => {
-        if (!orgId && !cardId) return;
+        if (!orgId || isNaN(orgId) || !cardId || isNaN(cardId)) return;
 
         alert.destory({
             title: '카드를 삭제하시겠습니까?',
             confirmFn: () => creditCardApi.destroy(orgId, cardId),
-            routerFn: () => router.replace(V3OrgCardListPageRoute.path(orgId)),
+            routerFn: () => {
+                router.replace(V3OrgCardListPageRoute.path(orgId));
+                setCardId(null);
+            },
         });
     };
